+++ packages/services/f-statistics/src/services/_tests/statistics.service.test.js
@@ -59,7 +59,7 @@ describe('f-statistics', () => {
             };
 
             // Act
-            const statisticsService = new StatisticsService();
+            const statisticsService = new StatisticsService(null, null, null, store);
 
             // Assert
             expect(statisticsService.basePayload).toEqual(expectedPayload);