+++ packages/services/f-statistics/src/store/logic/_tests/startBatchPublishTimer.test.js
@@ -1,0 +1,13 @@
+// import startBatchPublishTimer from '../startBatchPublishTimer';
+
+describe('Statistics Store Logic ::', () => {
+    afterEach(() => {
+        jest.clearAllMocks();
+    });
+
+    describe.skip('startBatchPublishTimer ::', () => {
+        it.skip('should do something', () => {
+
+        });
+    });
+});